Fix Lb302 silence This problem seem to arise due to casting _n->framesLeft() and _n->offset() from an unsigned type (size_t) to a signed type (int). The fix is to use size_t as the type for std::max across the board.

+++ plugins/Lb302/Lb302.cpp
@@ -751,7 +751,7 @@ void Lb302Synth::playNote( NotePlayHandle * _n, SampleFrame* _working_buffer )
 	}
 	m_notesMutex.unlock();
 
-	release_frame = std::max(release_frame, static_cast<int>(_n->framesLeft()) + static_cast<int>(_n->offset()));
+	release_frame = std::max(release_frame, _n->framesLeft() + _n->offset());
 }
